+++ build.sh
@@ -13,4 +13,6 @@ if [ ! -d build ]; then
     mkdir build
 fi
 
-"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/signals.f90 app/main.f90 -o build/forsynth.out
+rm *.mod
+
+"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/signals.f90 src/music.f90 src/audio_effects.f90 src/demos.f90 app/main.f90 -o build/forsynth.out