+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to the Forsynth project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+
+## [Forsynth development version]
+
+### Added
+- music module:
+    - `add_note()` to generate a note with a sum of harmonics.
+    - `add_major_chord()`
+    - `add_minor_chord()`
+
+
 ## [Forsynth 0.1 "Stockhausen"] 2021-02-19
 
 ### Added